src/cloud_functions/vertex-at-stage1-power-prediction-invoker-XAI/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
google-cloud-pubsub